+++ src/frontend-cdoc-commons/components/cdoc-odobjectdetails/cdoc-odobjectdetails.component.ts
@@ -1,8 +1,6 @@
 import {ChangeDetectionStrategy, ChangeDetectorRef, Component, Input, OnInit} from '@angular/core';
 import {AbstractInlineComponent} from '../../../angular-commons/components/inline.component';
-import {
-    BaseObjectDetectionImageObjectRecordType
-} from '@dps/mycms-commons/dist/search-commons/model/records/baseobjectdetectionimageobject-record';
+import {BaseObjectDetectionImageObjectRecordType} from '@dps/mycms-commons/dist/search-commons/model/records/baseobjectdetectionimageobject-record';
 import {AppState, GenericAppService} from '@dps/mycms-commons/dist/commons/services/generic-app.service';
 import {BeanUtils} from '@dps/mycms-commons/dist/commons/utils/bean.utils';