+++ src/frontend-cdoc-commons/resolver/cdoc-section-searchform.resolver.ts
@@ -3,9 +3,7 @@ import {AppState, GenericAppService} from '@dps/mycms-commons/dist/commons/servi
 import {ResolvedData, ResolverError} from '../../angular-commons/resolver/resolver.utils';
 import {IdValidationRule} from '@dps/mycms-commons/dist/search-commons/model/forms/generic-validator.util';
 import {CommonDocSearchForm} from '@dps/mycms-commons/dist/search-commons/model/forms/cdoc-searchform';
-import {
-    GenericSearchFormSearchFormConverter
-} from '@dps/mycms-commons/dist/search-commons/services/generic-searchform.converter';
+import {GenericSearchFormSearchFormConverter} from '@dps/mycms-commons/dist/search-commons/services/generic-searchform.converter';
 
 export abstract class CommonSectionSearchFormResolver<F extends CommonDocSearchForm> implements Resolve<ResolvedData<F>> {
     static ERROR_INVALID_SEARCHFORM_SECTION_ID = 'ERROR_INVALID_SEARCHFORM_SECTION_ID';