+++ src/frontend-section-commons/components/sectionpage/section-page.component.spec.ts
@@ -12,7 +12,6 @@ import {StaticPagesDataService} from '@dps/mycms-commons/dist/pdoc-commons/servi
 import {CommonDocRoutingService} from '../../../frontend-cdoc-commons/services/cdoc-routing.service';
 import {ErrorResolver} from '../../../frontend-cdoc-commons/resolver/error.resolver';
 import {PageUtils} from '../../../angular-commons/services/page.utils';
-import {AngularMarkdownService} from '../../../angular-commons/services/angular-markdown.service';
 import {AngularHtmlService} from '../../../angular-commons/services/angular-html.service';
 import {CommonRoutingService} from '../../../angular-commons/services/common-routing.service';
 import {RouterStub} from '../../../angular-commons/testing/router-stubs';
@@ -24,6 +23,9 @@ import {LayoutService} from '../../../angular-commons/services/layout.service';
 import {AppServiceStub} from '../../../angular-commons/testing/appservice-stubs';
 import {GenericAppService} from '@dps/mycms-commons/dist/commons/services/generic-app.service';
 import {ToastrServiceStub} from '../../../testing/toasts-stubs';
+import {SimpleAngularMarkdownService} from '../../../angular-commons/services/simple-angular-markdown.service';
+import {AngularMarkdownService} from '../../../angular-commons/services/angular-markdown.service';
+import {SimpleAngularHtmlService} from '../../../angular-commons/services/simple-angular-html.service';
 
 describe('SectionPageComponent', () => {
     let component: SectionPageComponent;
@@ -46,8 +48,8 @@ describe('SectionPageComponent', () => {
                 SearchParameterUtils,
                 { provide: ToastrService, useValue: new ToastrServiceStub() },
                 TranslateService,
-                AngularMarkdownService,
-                AngularHtmlService,
+                {provide: AngularMarkdownService, useClass: SimpleAngularMarkdownService},
+                {provide: AngularHtmlService, useClass: SimpleAngularHtmlService},
                 ErrorResolver,
                 PageUtils,
                 GenericTrackingService,