+++ src/frontend-section-module/sections.module.ts
@@ -7,8 +7,6 @@ import {FormsModule, ReactiveFormsModule} from '@angular/forms';
 import {SectionsPDocsResolver} from '../frontend-cdoc-commons/resolver/sections-pdocs.resolver';
 import {ErrorResolver} from '../frontend-cdoc-commons/resolver/error.resolver';
 import {PageUtils} from '../angular-commons/services/page.utils';
-import {AngularMarkdownService} from '../angular-commons/services/angular-markdown.service';
-import {AngularHtmlService} from '../angular-commons/services/angular-html.service';
 import {CommonRoutingService} from '../angular-commons/services/common-routing.service';
 import {TranslateModule} from '@ngx-translate/core';
 import {FrontendSectionCommonsModule} from '../frontend-section-commons/frontend-section-commons.module';
@@ -32,9 +30,8 @@ import {FrontendCommonDocCommonsModule} from '../frontend-cdoc-commons/frontend-
         SectionsPDocRecordResolver,
         SectionsPDocsResolver,
         ErrorResolver,
-        AngularHtmlService,
-        AngularMarkdownService,
         PageUtils
     ]
 })
-export class SectionsModule {}
+export class SectionsModule {
+}